improved picture trigger

+++ server and utilities/server-data.py
@@ -18,37 +18,42 @@ def main():
     objects = server.get_objects_node()
     opc_db = objects.add_object(idx, "OPC Testing")
 
-
     # Add tags
-    
+
     # depth
-    roi_depth_node = opc_db.add_variable(idx, "roi_depth_node", 0, ua.VariantType.Float)
+    roi_depth_node = opc_db.add_variable(
+        idx, "roi_depth_node", 0, ua.VariantType.Float)
     roi_depth_node.set_writable(writable=True)
-    
+
     # accuracy
-    roi_accuracy_node = opc_db.add_variable(idx, "roi_accuracy_node", 0, ua.VariantType.Float)
+    roi_accuracy_node = opc_db.add_variable(
+        idx, "roi_accuracy_node", 0, ua.VariantType.Float)
     roi_accuracy_node.set_writable(writable=True)
-    
+
     # roi select
-    roi_select_node = opc_db.add_variable(idx, "roi_select_node", 0, ua.VariantType.Float)
+    roi_select_node = opc_db.add_variable(
+        idx, "roi_select_node", 0, ua.VariantType.Int16)
     roi_select_node.set_writable(writable=True)
 
     # status
-    status_node = opc_db.add_variable(idx, "status_node", 0, ua.VariantType.Float)
-    status_node.set_writable(writable=True)    
+    status_node = opc_db.add_variable(
+        idx, "status_node", 0, ua.VariantType.Int16)
+    status_node.set_writable(writable=True)
 
     # trigger
-    picture_trigger_node = opc_db.add_variable(idx, "picture_trigger_node", 0, ua.VariantType.Boolean)
+    picture_trigger_node = opc_db.add_variable(
+        idx, "picture_trigger_node", 0, ua.VariantType.Boolean)
     picture_trigger_node.set_writable(writable=True)
-    
+
     # alive
-    alive_node = opc_db.add_variable(idx, "alive_node", 0, ua.VariantType.Boolean)
+    alive_node = opc_db.add_variable(
+        idx, "alive_node", 0, ua.VariantType.Boolean)
     alive_node.set_writable(writable=True)
-    
+
     dead_count = 0
     picture_count = 0
-    
-    while True:   
+
+    while True:
         # get tag values
         depth = roi_depth_node.get_value()
         accuracy = roi_accuracy_node.get_value()
@@ -56,9 +61,9 @@ def main():
         picture = picture_trigger_node.get_value()
         alive = alive_node.get_value()
         roi_select = roi_select_node.get_value()
-        
+
         picture_count += 1
-        
+
         if picture_count == 1:
             dv = True
             dv = ua.DataValue(ua.Variant(dv, ua.VariantType.Boolean))
@@ -69,7 +74,7 @@ def main():
             picture_trigger_node.set_value(dv)
         if picture_count > 20:
             picture_count = 0
-        
+
         # check if client is alive
         if alive:
             dead_count = 0
@@ -78,19 +83,18 @@ def main():
             alive_node.set_value(dv)
         else:
             dead_count += 1
-        
+
         # print tag values
         print(f'Depth: {depth:.2f} | Accuracy: {accuracy:.2f} | '
               f'Status: {status:.2f} | Picture: {picture:.2f} | '
               f'Alive: {alive:.2f} | ROI select: {roi_select:.2f} | '
               f'Dead: {dead_count} | PicCount: {picture_count}')
-        
+
         time.sleep(0.5)
 
 
 if __name__ == "__main__":
     main()
-
 
     # old stuff
 
@@ -112,8 +116,7 @@ if __name__ == "__main__":
     #     idx, "Realsense Depth Array", 0, ua.VariantType.Float)
     # array_tag.set_writable(writable=True)
     # array_tag.set_value(0)
-    
-    
+
     # depth = depth_tag.get_value()
     # timer = timer_tag.get_value()
     # tick = client_tick_tag.get_value()
